# docs/api-contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AshyqQala.kz API
  version: 0.1.0
  description: >
    Публичный read-only API. Рукописный АРБИТР контракта: openapi-typescript генерит TS-типы фронта,
    kin-openapi валидирует ответы Go-хендлеров в тестах (НЕ генерация из хендлеров).
    Честный конверт {value,state}: value:null ⇒ "no_data". Полный value_state-enum — из registry (Story 1.4).

paths:
  /healthz:
    get:
      summary: Health-проба (БД доступна)
      responses:
        '200':
          description: сервис и БД доступны
          content:
            text/plain:
              schema: { type: string }
        '503':
          description: БД недоступна
          content:
            text/plain:
              schema: { type: string }

  /api/contracts/{goszakup_id}:
    get:
      summary: Карточка контракта по natural goszakup_id
      parameters:
        - name: goszakup_id
          in: path
          required: true
          description: Публичный natural id (НЕ суррогатный bigint)
          schema: { type: string }
      responses:
        '200':
          description: контракт найден
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Contract' }
        '404':
          description: контракт не найден
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          description: внутренняя ошибка
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /api/contractors/{bin}:
    get:
      summary: Карточка подрядчика по natural БИН (Story 5.2, FR-13/FR-14)
      parameters:
        - name: bin
          in: path
          required: true
          description: Публичный natural БИН организации (12 цифр)
          schema: { type: string }
      responses:
        '200':
          description: подрядчик найден
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Contractor' }
        '404':
          description: подрядчик не найден
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          description: внутренняя ошибка
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /api/contracts/{goszakup_id}/flags/{flag_type}/evidence.json:
    get:
      summary: Экспорт evidence raised-флага — канонический JSON (Story 5.6, AR-29/SM-5)
      description: >
        Пересчитываемый третьим лицом/СМИ экспорт evidence ОДНОГО raised-флага контракта. subject_ref —
        публичный goszakup_contract_id (не суррогатный bigint). Не-raised флаг / отсутствие → 404 (цитировать нечего).
        Параллельный печатный нейтральный текст — .txt (text/plain, вне JSON-арбитра, как OG-HTML Story 5.5).
      parameters:
        - name: goszakup_id
          in: path
          required: true
          schema: { type: string }
        - name: flag_type
          in: path
          required: true
          schema: { type: string, enum: [single_participant, price_per_km] }
        - name: lang
          in: query
          required: false
          description: язык нейтральной рамки печатного .txt; JSON локаль-агностичен
          schema: { type: string, enum: [ru, kk] }
      responses:
        '200':
          description: канонический экспорт evidence
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EvidenceExport' }
        '404':
          description: нет raised-флага этого типа у контракта (цитировать нечего)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          description: внутренняя ошибка
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /api/methodology:
    get:
      summary: Пороги методики (Story 5.3, FR-23) — единый источник для экрана методики
      description: >
        Read-only пороги из methodology_params (registry). Экран методики показывает формулу/пороги ВСЕГДА
        (в т.ч. когда сигнал не выставлен), без литералов на фронте. Иммутабельно в рантайме.
      responses:
        '200':
          description: версия + пороги методики
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Methodology' }
        '500':
          description: внутренняя ошибка
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /api/error-reports:
    post:
      summary: Канал «сообщить об ошибке» (Story 5.4, FR-28) — публичный безаккаунтный досудебный канал
      description: >
        ПЕРВЫЙ write-эндпоинт. Принимает обращение об ошибке в данных/флаге/геопривязке и кладёт в очередь
        error_reports (Directus триажит). Защита: honeypot (скрытое поле leave_blank) + лимит тела + rate-limit per-IP.
        Без аккаунта, без публичного трекинга статуса (PRD §5.11).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ErrorReportRequest' }
      responses:
        '201':
          description: обращение принято в очередь
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorReportResponse' }
        '400':
          description: некорректный вход
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429':
          description: слишком много обращений (rate-limit)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          description: внутренняя ошибка
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /api/lots:
    get:
      summary: ⏳ Лоты Астаны с интерим-гео для ранней карты (Story 0.8, трек «Парсер-мост»)
      description: >
        Временный read-only список scraped-лотов (проекция lots + интерим-гео interim_geo_lots) для
        ранней карты. ТОЛЬКО лоты — карточки контрактов, флаги, медианы ждут токен ows_v2. Координаты —
        bare [lon,lat] (nullable; null ⇒ нет точки, НЕ 0,0). Замещается живым импортом при swap scrape→ows.
      responses:
        '200':
          description: список лотов (точки + честные состояния геопривязки)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MapLot' }
        '500':
          description: внутренняя ошибка
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

components:
  schemas:
    StringField:
      type: object
      additionalProperties: false
      required: [value, state]
      properties:
        value:
          type: string
          nullable: true
        state:
          type: string
          # value_state — синхронно с registry/values/honest_states.json (перекрёстный тест, Story 1.4 AC4)
          enum: [ok, no_data, insufficient_sample, not_comparable, stale, geocode_pending, geocode_failed, source_conflict, redacted, not_applicable, error]
    FlagField:
      type: object
      additionalProperties: false
      # Задел карточки флага (Epic 4/5): двухосевой enum, ось flag_state. Эндпоинты появятся позже.
      required: [state]
      properties:
        state:
          type: string
          # flag_state — синхронно с registry/values/honest_states.json (перекрёстный тест, Story 1.4 AC4)
          enum: [raised, not_raised, insufficient_data, not_published]
    Act:
      type: object
      additionalProperties: false
      # Акт приёмки на карточке (FR-11, Story 5.1). present — есть ли акт; при отсутствии поля no_data
      # (карточка не выдумывает дату/подписанта). signer — служебная информация должностного лица (§6.2.4).
      required: [present, act_date, signer, source_url]
      properties:
        present: { type: boolean }
        act_date: { $ref: '#/components/schemas/StringField' }
        signer: { $ref: '#/components/schemas/StringField' }
        source_url: { $ref: '#/components/schemas/StringField' }
    ContractFlag:
      type: object
      additionalProperties: false
      # Флаг на карточке контракта (FR-12/FR-23, Story 5.1). Контракт-субъектные флаги (FR-19/FR-20).
      # state — ось flag_state honest-enum, реконструированная на чтении (AC4): нет строки → insufficient_data.
      # evidence — СЫРОЙ jsonb пересчитываемости, непусто ТОЛЬКО при raised (иначе null). methodology_version —
      # value лишь когда флаг оценивался (есть строка), иначе no_data.
      required: [flag_id, state, methodology_version, detected_at, evidence]
      properties:
        flag_id:
          type: string
          enum: [single_participant, price_per_km]
        state:
          type: string
          enum: [raised, not_raised, insufficient_data, not_published]
        methodology_version: { $ref: '#/components/schemas/StringField' }
        detected_at: { $ref: '#/components/schemas/StringField' } # когда оценён (ISO8601 Z); no_data если строки нет
        evidence:
          type: object
          nullable: true
          additionalProperties: true
    MethodologyDrift:
      type: object
      additionalProperties: false
      # AR-29 (Story 5.6): честный сигнал, что перманентная ссылка-на-дату создана под ИНОЙ версией методики,
      # чем текущая. present=false при отсутствии дрейфа/штампа. Карточка ВСЕГДА показывает текущее состояние;
      # исторический срез в MVP не хранится (нет snapshot_id) — дрейф лишь честно сообщает «методика сменилась».
      required: [present, requested_version, current_version]
      properties:
        present: { type: boolean }
        requested_version: { type: string } # mv из ссылки (пусто, если штампа не было)
        current_version: { type: string }   # текущая каноническая methodology_version
    EvidenceExport:
      type: object
      additionalProperties: false
      # AR-29 (Story 5.6): канонический экспорт evidence raised-флага для цитирования третьим лицом/СМИ.
      # subject_ref — ПУБЛИЧНЫЙ goszakup_contract_id (НЕ внутренний bigint). evidence — сырой пересчитываемый объект.
      required: [flag_type, subject_type, subject_ref, methodology_version, evidence]
      properties:
        flag_type: { type: string, enum: [single_participant, price_per_km] }
        subject_type: { type: string }
        subject_ref: { type: string }
        methodology_version: { type: string }
        evidence:
          type: object
          additionalProperties: true
    Contract:
      type: object
      additionalProperties: false
      required:
        - goszakup_contract_id
        - subject_ru
        - subject_kk
        - amount_tng
        - sign_date
        - plan_start
        - plan_end
        - status
        - direction
        - kato_code
        - customer
        - supplier
        - source_url
        - act
        - flags
        - imported_at
        - updated_at
        - methodology_version
        - as_of
        - methodology_drift
      properties:
        goszakup_contract_id: { type: string }
        subject_ru: { $ref: '#/components/schemas/StringField' }
        subject_kk: { $ref: '#/components/schemas/StringField' }
        amount_tng: { $ref: '#/components/schemas/StringField' } # целые тенге СТРОКОЙ
        sign_date: { $ref: '#/components/schemas/StringField' }
        plan_start: { $ref: '#/components/schemas/StringField' }
        plan_end: { $ref: '#/components/schemas/StringField' }
        status: { $ref: '#/components/schemas/StringField' }
        direction: { $ref: '#/components/schemas/StringField' }
        kato_code: { $ref: '#/components/schemas/StringField' }
        customer: { $ref: '#/components/schemas/StringField' } # Заказчик — no_data до Epic 2 (organizations)
        supplier: { $ref: '#/components/schemas/StringField' } # Подрядчик — no_data до Epic 2
        source_url: { $ref: '#/components/schemas/StringField' }
        act: { $ref: '#/components/schemas/Act' }                  # FR-11: акт приёмки (всегда объект; present=false если нет)
        flags:                                                    # FR-12/FR-23: контракт-флаги + честная реконструкция
          type: array
          items: { $ref: '#/components/schemas/ContractFlag' }
        imported_at: { $ref: '#/components/schemas/StringField' } # ISO8601 Z в value; единый конверт + честный no_data
        updated_at: { $ref: '#/components/schemas/StringField' }
        # AR-29 (Story 5.6) — перманентная ссылка-на-дату:
        methodology_version: { $ref: '#/components/schemas/StringField' } # текущая каноническая версия (штамп ссылки/дисплей)
        as_of: { $ref: '#/components/schemas/StringField' }               # дата-штамп ссылки (detected_at raised / updated_at)
        methodology_drift: { $ref: '#/components/schemas/MethodologyDrift' } # present=true если ссылка под иной версией методики
    ContractorProfile:
      type: object
      additionalProperties: false
      # Честное состояние профиля подрядчика (FR-13, Story 5.2). incomplete — связь контракт↔орг ещё не
      # наполнена живым импортом (Story 2.2) → «профиль неполный». unverified — у орг есть manual/conflict
      # псевдонимы → «профиль уточняется», флаги НЕ основание. "complete" в S-0 не выдаётся.
      required: [state]
      properties:
        state: { type: string, enum: [incomplete, partial, unverified] }
    ContractorContract:
      type: object
      additionalProperties: false
      # Элемент списка контрактов подрядчика (FR-13, Story 5.2). goszakup_contract_id — публичная ссылка на карточку.
      required: [goszakup_contract_id, subject_ru, subject_kk, amount_tng, kato_code, direction]
      properties:
        goszakup_contract_id: { type: string }
        subject_ru: { $ref: '#/components/schemas/StringField' }
        subject_kk: { $ref: '#/components/schemas/StringField' }
        amount_tng: { $ref: '#/components/schemas/StringField' }
        kato_code: { $ref: '#/components/schemas/StringField' }
        direction: { $ref: '#/components/schemas/StringField' }
    ContractorFlag:
      type: object
      additionalProperties: false
      # Агрегированный флаг подрядчика (FR-21 монополия / FR-22 РНУ, Story 5.2). state — честная реконструкция
      # на чтении; при unverified-профиле принудительно insufficient_data (флаг не основание).
      required: [flag_id, state, methodology_version, detected_at, evidence]
      properties:
        flag_id: { type: string, enum: [monopoly, rnu] }
        state: { type: string, enum: [raised, not_raised, insufficient_data, not_published] }
        methodology_version: { $ref: '#/components/schemas/StringField' }
        detected_at: { $ref: '#/components/schemas/StringField' }
        evidence:
          type: object
          nullable: true
          additionalProperties: true
    RNUMark:
      type: object
      additionalProperties: false
      # Метка РНУ (FR-14, Story 5.2). active реконструируется на чтении из дат (авто-снятие по end_date).
      # У КАЖДОЙ метки свой source_url (AC-3: защитная ссылка на каждую метку). Нейтральные данные only.
      required: [active, start_date, end_date, reason_ref, source_url, registry_id]
      properties:
        active: { type: boolean }
        start_date: { $ref: '#/components/schemas/StringField' }
        end_date: { $ref: '#/components/schemas/StringField' }
        reason_ref: { $ref: '#/components/schemas/StringField' }
        source_url: { $ref: '#/components/schemas/StringField' }
        registry_id: { $ref: '#/components/schemas/StringField' }
    Contractor:
      type: object
      additionalProperties: false
      # Карточка подрядчика (FR-13/FR-14, Story 5.2). Публичный id = natural bin. Агрегаты count/amount —
      # no_data при неполном профиле (linkage до Story 2.2), НЕ «0». regions — distinct КАТО.
      required:
        - bin
        - name_ru
        - name_kk
        - reg_kato
        - is_customer
        - is_supplier
        - profile
        - contract_count
        - total_amount_tng
        - regions
        - contracts
        - flags
        - rnu_marks
        - source_url
        - imported_at
        - updated_at
      properties:
        bin: { type: string }
        name_ru: { $ref: '#/components/schemas/StringField' }
        name_kk: { $ref: '#/components/schemas/StringField' }
        reg_kato: { $ref: '#/components/schemas/StringField' }
        is_customer: { type: boolean }
        is_supplier: { type: boolean }
        profile: { $ref: '#/components/schemas/ContractorProfile' }
        contract_count: { $ref: '#/components/schemas/StringField' }   # целое строкой; no_data при неполном профиле
        total_amount_tng: { $ref: '#/components/schemas/StringField' } # тенге строкой; no_data при неполном профиле
        regions:
          type: array
          items: { type: string }
        contracts:
          type: array
          items: { $ref: '#/components/schemas/ContractorContract' }
        flags:
          type: array
          items: { $ref: '#/components/schemas/ContractorFlag' }
        rnu_marks:
          type: array
          items: { $ref: '#/components/schemas/RNUMark' }
        source_url: { $ref: '#/components/schemas/StringField' }
        imported_at: { $ref: '#/components/schemas/StringField' }
        updated_at: { $ref: '#/components/schemas/StringField' }
    MapLot:
      type: object
      additionalProperties: false
      # ⏳ ИНТЕРИМ (Story 0.8, трек «Парсер-мост»). Лот на ранней карте: наименование/сумма — честный
      # конверт {value,state}; координаты — bare [lon,lat] (null ⇒ нет точки, НЕ 0,0).
      description: >-
        Инвариант координат (OpenAPI 3.0 не выражает условную required-зависимость декларативно,
        поэтому он задан здесь как контракт): geocode_state = ok ГАРАНТИРУЕТ, что lon и lat
        не null (точка на карте есть); при ЛЮБОМ другом состоянии (geocode_pending, geocode_failed)
        lon и lat равны null. Координаты НИКОГДА не равны 0,0.
      required:
        - goszakup_lot_id
        - subject_ru
        - subject_kk
        - amount_tng
        - lon
        - lat
        - geocode_state
      properties:
        goszakup_lot_id: { type: string }
        subject_ru: { $ref: '#/components/schemas/StringField' }
        subject_kk: { $ref: '#/components/schemas/StringField' }
        amount_tng: { $ref: '#/components/schemas/StringField' } # целые тенге СТРОКОЙ
        lon: { type: number, nullable: true } # GeoJSON [lon,lat]; null когда нет точки (НЕ 0,0)
        lat: { type: number, nullable: true }
        geocode_state:
          type: string
          # Подмножество value_state (ось 1): почему лот на карте/вне карты. ok = есть точка;
          # geocode_pending = ещё не геокодирован (нет гео-строки); geocode_failed = unmatched (без точки).
          enum: [ok, geocode_pending, geocode_failed]
    MethodologyThresholds:
      type: object
      additionalProperties: false
      # Пороги синхронны с methodology_params.v1.yaml (FR-23). Доли/факторы — числа (малые десятичные).
      required:
        - min_sample
        - comparability_window_months
        - price_per_km_deviation_factor
        - monopoly_concentration_share
        - monopoly_min_group_contracts
        - single_participant_exclude_methods
      properties:
        min_sample: { type: integer }
        comparability_window_months: { type: integer }
        price_per_km_deviation_factor: { type: number }
        monopoly_concentration_share: { type: number }
        monopoly_min_group_contracts: { type: integer }
        single_participant_exclude_methods:
          type: array
          items: { type: string }
    Methodology:
      type: object
      additionalProperties: false
      required: [version, thresholds]
      properties:
        version: { type: string } # каноническая methodology_version (^v\d+\.\d+$)
        thresholds: { $ref: '#/components/schemas/MethodologyThresholds' }
    ErrorReportRequest:
      type: object
      additionalProperties: false
      # Публичное обращение (FR-28, Story 5.4). kind/subject_type — доменные enum (зеркало CHECK миграции 0011).
      # leave_blank — honeypot (скрытое поле, неавтозаполняемое имя; должно быть пустым). contact/source_url опц.
      required: [kind, subject_type, subject_ref, message]
      properties:
        kind: { type: string, enum: [data_error, flag_error, geo_wrong_point] }
        subject_type: { type: string, enum: [contract, contractor, geo_object] }
        subject_ref: { type: string, minLength: 1, maxLength: 128 }
        message: { type: string, minLength: 1, maxLength: 5000 }
        contact: { type: string, maxLength: 200 }
        source_url: { type: string, maxLength: 512 }
        leave_blank: { type: string } # honeypot — должен быть пустым
    ErrorReportResponse:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        id: { type: integer, format: int64 } # отсутствует при honeypot-приёме
        status: { type: string, enum: [received] }
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              enum: [NOT_FOUND, VALIDATION_FAILED, INVALID_CURSOR, RATE_LIMITED, INTERNAL]
            message: { type: string }