# server/internal/store/gen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.0

package gen

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Act struct {
	ID            int64              `json:"id"`
	ContractID    int64              `json:"contract_id"`
	GoszakupActID string             `json:"goszakup_act_id"`
	ActDate       pgtype.Date        `json:"act_date"`
	SignerInfo    pgtype.Text        `json:"signer_info"`
	SourceUrl     pgtype.Text        `json:"source_url"`
	ImportedAt    pgtype.Timestamptz `json:"imported_at"`
	UpdatedAt     pgtype.Timestamptz `json:"updated_at"`
}

type Contract struct {
	ID                 int64              `json:"id"`
	GoszakupContractID string             `json:"goszakup_contract_id"`
	SubjectRu          pgtype.Text        `json:"subject_ru"`
	SubjectKk          pgtype.Text        `json:"subject_kk"`
	AmountTng          pgtype.Int8        `json:"amount_tng"`
	SignDate           pgtype.Date        `json:"sign_date"`
	PlanStart          pgtype.Date        `json:"plan_start"`
	PlanEnd            pgtype.Date        `json:"plan_end"`
	Status             pgtype.Text        `json:"status"`
	Direction          pgtype.Text        `json:"direction"`
	KatoCode           pgtype.Text        `json:"kato_code"`
	SourceUrl          pgtype.Text        `json:"source_url"`
	IsDeleted          bool               `json:"is_deleted"`
	ImportedAt         pgtype.Timestamptz `json:"imported_at"`
	UpdatedAt          pgtype.Timestamptz `json:"updated_at"`
	CustomerOrgID      pgtype.Int8        `json:"customer_org_id"`
	SupplierOrgID      pgtype.Int8        `json:"supplier_org_id"`
}

type ErrorReport struct {
	ID          int64              `json:"id"`
	Kind        string             `json:"kind"`
	SubjectType string             `json:"subject_type"`
	SubjectRef  string             `json:"subject_ref"`
	Message     string             `json:"message"`
	Contact     pgtype.Text        `json:"contact"`
	SourceUrl   pgtype.Text        `json:"source_url"`
	Status      string             `json:"status"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
	ResolvedAt  pgtype.Timestamptz `json:"resolved_at"`
}

type FlagDispute struct {
	ID          int64              `json:"id"`
	RiskFlagID  int64              `json:"risk_flag_id"`
	FlagType    string             `json:"flag_type"`
	SubjectType string             `json:"subject_type"`
	SubjectID   int64              `json:"subject_id"`
	Status      string             `json:"status"`
	Note        pgtype.Text        `json:"note"`
	SourceUrl   pgtype.Text        `json:"source_url"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
	ResolvedAt  pgtype.Timestamptz `json:"resolved_at"`
}

type InterimGeoLot struct {
	ID            int64              `json:"id"`
	GoszakupLotID string             `json:"goszakup_lot_id"`
	Lat           pgtype.Float8      `json:"lat"`
	Lon           pgtype.Float8      `json:"lon"`
	KatoCode      pgtype.Text        `json:"kato_code"`
	GeocodeStatus string             `json:"geocode_status"`
	Confidence    pgtype.Float8      `json:"confidence"`
	AddressText   pgtype.Text        `json:"address_text"`
	GeocodedAt    pgtype.Timestamptz `json:"geocoded_at"`
}

type Lot struct {
	ID             int64              `json:"id"`
	GoszakupLotID  string             `json:"goszakup_lot_id"`
	AnnouncementID pgtype.Int8        `json:"announcement_id"`
	TitleRu        pgtype.Text        `json:"title_ru"`
	TitleKk        pgtype.Text        `json:"title_kk"`
	Amount         pgtype.Int8        `json:"amount"`
	Quantity       pgtype.Int8        `json:"quantity"`
	Unit           pgtype.Text        `json:"unit"`
	KatoCode       pgtype.Text        `json:"kato_code"`
	IsDeleted      bool               `json:"is_deleted"`
	ImportedAt     pgtype.Timestamptz `json:"imported_at"`
	UpdatedAt      pgtype.Timestamptz `json:"updated_at"`
}

type MethodologyParam struct {
	ID            int64              `json:"id"`
	Version       string             `json:"version"`
	Key           string             `json:"key"`
	Value         string             `json:"value"`
	Description   pgtype.Text        `json:"description"`
	EffectiveFrom pgtype.Timestamptz `json:"effective_from"`
}

type OrgNameAlias struct {
	ID             int64              `json:"id"`
	OrganizationID pgtype.Int8        `json:"organization_id"`
	RawName        string             `json:"raw_name"`
	Source         string             `json:"source"`
	ResolveStatus  string             `json:"resolve_status"`
	DetectedAt     pgtype.Timestamptz `json:"detected_at"`
	UpdatedAt      pgtype.Timestamptz `json:"updated_at"`
}

type Organization struct {
	ID          int64              `json:"id"`
	Bin         string             `json:"bin"`
	NameRu      pgtype.Text        `json:"name_ru"`
	NameKk      pgtype.Text        `json:"name_kk"`
	RegKato     pgtype.Text        `json:"reg_kato"`
	IsCustomer  bool               `json:"is_customer"`
	IsSupplier  bool               `json:"is_supplier"`
	FirstSeenAt pgtype.Timestamptz `json:"first_seen_at"`
	SourceUrl   pgtype.Text        `json:"source_url"`
	IsDeleted   bool               `json:"is_deleted"`
	ImportedAt  pgtype.Timestamptz `json:"imported_at"`
	UpdatedAt   pgtype.Timestamptz `json:"updated_at"`
}

type PriceBenchmark struct {
	ID                 int64              `json:"id"`
	ComparabilityKey   string             `json:"comparability_key"`
	MedianPricePerKm   pgtype.Int8        `json:"median_price_per_km"`
	SampleSize         int32              `json:"sample_size"`
	MethodologyVersion string             `json:"methodology_version"`
	ComputedAt         pgtype.Timestamptz `json:"computed_at"`
}

type RiskFlag struct {
	ID                 int64              `json:"id"`
	FlagType           string             `json:"flag_type"`
	SubjectType        string             `json:"subject_type"`
	ContractID         pgtype.Int8        `json:"contract_id"`
	OrganizationID     pgtype.Int8        `json:"organization_id"`
	Severity           pgtype.Text        `json:"severity"`
	Evidence           []byte             `json:"evidence"`
	IsActive           bool               `json:"is_active"`
	DetectedAt         pgtype.Timestamptz `json:"detected_at"`
	ClearedAt          pgtype.Timestamptz `json:"cleared_at"`
	MethodologyVersion string             `json:"methodology_version"`
}

type RnuEntry struct {
	ID             int64       `json:"id"`
	OrganizationID int64       `json:"organization_id"`
	GoszakupRnuID  pgtype.Text `json:"goszakup_rnu_id"`
	StartDate      pgtype.Date `json:"start_date"`
	EndDate        pgtype.Date `json:"end_date"`
	ReasonRef      pgtype.Text `json:"reason_ref"`
	SourceUrl      pgtype.Text `json:"source_url"`
}